% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{calc_modularity}
\alias{calc_modularity}
\title{Calculate Adjusted Modularity Score}
\usage{
calc_modularity(feat_mat, labels, digits = 3, knn_k = NULL)
}
\arguments{
\item{feat_mat}{Numeric matrix or data frame. The feature matrix used to
compute the SNN graph (e.g., CLR abundances).}

\item{labels}{Vector of factors or character strings. The cluster assignments
for each row in \code{feat_mat}.}

\item{digits}{Integer (default: \code{3}). The number of decimal places to
round the final adjusted score.}

\item{knn_k}{Integer (optional, default: \code{NULL}). The number of nearest
neighbors (\code{k}) used for SNN graph construction. If \code{NULL}, it
defaults to \code{max(3, round(sqrt(N)))}, where \code{N} is the number of
samples.}
}
\value{
A numeric value representing the adjusted modularity score.
}
\description{
Calculates the Modularity score for a given clustering (\code{labels}) based
on a Shared Nearest Neighbor (SNN) graph constructed from the feature matrix
(\code{feat_mat}). The score is adjusted by the theoretical maximum
modularity for the number of groups to always be between
-0.5 (poor clustering) and +1 (excellent clustering)).
}
