% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{data_gong_sharma}
\alias{data_gong_sharma}
\title{Gong & Sharma et al. dataset}
\format{
Data frame of cell type counts at high resolution annotation (samples as rows, cell types as columns).
}
\source{
The original data is from the study: \doi{10.1101/2024.09.10.612119}.
}
\usage{
data_gong_sharma
}
\description{
The original data from Gong & Sharma et al. sound_life_AIFI_L3_frequencies.csv file.
}
\keyword{datasets}
