% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{get_ecoda_assay}
\alias{get_ecoda_assay}
\title{Helper to get assay for ecoda SummarizedExperiment objects}
\usage{
get_ecoda_assay(se, assay)
}
\arguments{
\item{se}{A \link[=SummarizedExperiment-class]{SummarizedExperiment} object.}

\item{assay}{Character string, the name of the assay.}
}
\value{
A data frame containing the assay data.
}
\description{
Because not all data can be stored in assays (e.g. pseudobulk
and clr_hvc have different number of features (rows) than counts etc.)
}
