% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{plot_pca3d}
\alias{plot_pca3d}
\title{Plot 3-dimensional interactive Principal Component Analysis plot}
\usage{
plot_pca3d(
  ecoda_object,
  slot = c("clr", "clr_hvc", "counts", "counts_imp", "freq", "freq_imp", "asin_sqrt",
    "pb"),
  label_col = NULL,
  scale. = FALSE
)
}
\arguments{
\item{ecoda_object}{An \link[=ECODA-class]{ECODA} object.}

\item{slot}{Character string (default: \code{"clr"}). The name of the data
matrix slot in the \code{ECODA} object to use for PCA. Must be one of:
\code{"clr"} (CLR-transformed abundances, default), \code{"clr_hvc"}
(CLR-transformed abundances of only the most highly variable cell types
(HVCs)), \code{"pb"} (pseudobulk gene expression), \code{"counts"} (raw
counts), \code{"counts_imp"} (imputed counts), \code{"freq"} (relative
frequencies), \code{"freq_imp"} (imputed frequencies), or
\code{"asin_sqrt"} (arcsin-square root transformed data).}

\item{label_col}{Character string (optional, default: \code{NULL}). The name
of a column in \code{ecoda_object@metadata} used to color and group samples
in the plot, and for calculating clustering scores.}

\item{scale.}{Logical (default: \code{FALSE}). A value indicating whether the
variables should be scaled to have unit variance before the PCA.}
}
\value{
An interactive 3D \code{plotly} object visualizing the PCA results.
}
\description{
Performs Principal Component Analysis (PCA) on a selected data
matrix from the \code{ECODA} object (default: CLR-transformed abundances,
\code{clr}) and visualizes the results in 3D colored by \code{label_col}.
}
\examples{
data(example_data)
ecoda_object <- ecoda(
    data = example_data$Zhang$cell_counts_lowresolution,
    metadata = example_data$Zhang$metadata,
)

plot_pca3d(ecoda_object, label_col = "Tissue")
}
