% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.r
\name{compute_snn_graph}
\alias{compute_snn_graph}
\title{Compute Shared Nearest Neighbor (SNN) Graph}
\usage{
compute_snn_graph(knn)
}
\arguments{
\item{knn}{A matrix of nearest neighbor indices, typically generated by
\code{compute_KNN_from_dist}.}
}
\value{
An \code{igraph} graph object where edge weights correspond to the
count of shared neighbors between nodes.
}
\description{
Constructs a graph where nodes represent samples and edges are weighted
by the number of shared nearest neighbors (SNN) between them.
}
