% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scECODA-package.R
\docType{package}
\name{scECODA-package}
\alias{scECODA}
\alias{scECODA-package}
\title{scECODA: Single-Cell Exploratory Compositional Data Analysis}
\description{
The scECODA R package provides a complete workflow for the analysis and visualization of compositional data, primarily focusing on cell type proportions derived from single-cell data. It implements specialized methods, such as the Centered Log-Ratio (CLR) transformation, to properly analyze proportional data while avoiding the biases introduced by the compositional constraint. The package encapsulates data management, transformation, and analysis into a single SummarizedExperiment object, offering downstream tools for dimensionality reduction via PCA, calculating critical metrics like the Adjusted Rand Index (ARI) and Modularity to quantify sample grouping quality, and generating high-quality visualizations like heatmaps and scatter plots.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/carmonalab/scECODA}
  \item Report bugs at \url{https://github.com/carmonalab/scECODA/issues}
}

}
\author{
\strong{Maintainer}: Christian Halter \email{scecoda.dev@gmail.com} (\href{https://orcid.org/0009-0009-5479-2246}{ORCID})

Authors:
\itemize{
  \item Massimo Andreatta \email{Massimo.Andreatta@unige.ch} (\href{https://orcid.org/0000-0002-8036-2647}{ORCID})
  \item Santiago Carmona \email{Santiago.Carmona@unige.ch} (\href{https://orcid.org/0000-0002-2495-0671}{ORCID})
}

Other contributors:
\itemize{
  \item Swiss Cancer Research Foundation [funder]
}

}
\keyword{internal}
